requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.1
openai==1.30.1
python-multipart==0.0.9
PyMuPDF==1.23.22
faiss-cpu==1.7.4
numpy==1.26.4
tiktoken==0.6.0
